Moved the ParkingSlot list to a fragment (clearer and we'll be able to load different fragments dynamically later :) )
